refactor: extract editable logic from FormEditor into EditableContentBlock
