+++ src/explore-education-statistics-admin/src/utils/markdown/toHtml.ts
@@ -1,0 +1,5 @@
+import marked from 'marked';
+
+export default function toHtml(markdown: string) {
+  return marked(markdown);
+}
