+++ src/explore-education-statistics-admin/src/utils/markdown/toMarkdown.ts
@@ -1,0 +1,7 @@
+import Turndown from 'turndown';
+
+const turndownService = new Turndown();
+
+export default function toMarkdown(value: string) {
+  return turndownService.turndown(value);
+}
